rds-cluster-instance/requirements.txt
aws-cdk-lib==2.104.0
constructs>=10.1.83,<11.0.0
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-settings==2.0.3
pytest==7.1.2